clean up request handler
